# docs/openapi.yaml
openapi: 3.0.3
info:
  title: UNIPROM CRM API
  version: 1.0.0
  description: |
    HTTP API CRM для обработки заявок на изготовление изделий.

    Два взаимозаменяемых префикса с одинаковой серверной проверкой прав и области
    видимости: канонический `/api/v1/...` и контрактный `/api/v1/crm/...`.

    Аутентификация — JWT в заголовке `Authorization: Bearer <token>`.
    Технические эндпоинты дополнительно требуют заголовок `x-integration-token`.

    Создание заявки принимает заголовок `Idempotency-Key`. Изменение ресурса
    требует `lockVersion`; при конфликте возвращается 409 `VERSION_CONFLICT`.
  contact:
    name: UNIPROM
servers:
  - url: http://127.0.0.1:3001
    description: Локальный backend
...
paths:
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Вход в систему
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [login, password]
              properties:
                login: { type: string }
                password: { type: string, format: password }
      responses:
        '200':
          description: Токен и профиль
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '423': { $ref: '#/components/responses/Locked' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /api/v1/auth/me:
    get:
      tags: [auth]
      summary: Профиль текущего пользователя
      responses:
        '200':
          description: Профиль
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/password:
    post:
      tags: [auth]
      summary: Смена пароля
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string, format: password }
                newPassword: { type: string, format: password, minLength: 8 }
      responses:
        '200': { description: Пароль изменён }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/v1/applications:
    get:
      tags: [applications]
      summary: Список заявок в области видимости
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: stage
          in: query
          schema: { type: string }
        - name: complexity
          in: query
          schema: { type: string, enum: [STANDARD, MODIFIED, CUSTOM, NEEDS_CLASSIFICATION] }
        - name: ownerId
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Постраничный список
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PagedApplications' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [applications]
      summary: Создание заявки
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApplicationCreate' }
      responses:
        '201':
          description: Заявка создана
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApplicationCreateResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/applications/{number}:
    get:
      tags: [applications]
      summary: Карточка заявки
      parameters:
        - $ref: '#/components/parameters/Number'
      responses:
        '200':
          description: Заявка
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Application' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [applications]
      summary: Изменение заявки
      parameters:
        - $ref: '#/components/parameters/Number'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApplicationUpdate' }
      responses:
        '200':
          description: Заявка изменена
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Application' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/v1/applications/{number}/stage:
    post:
      tags: [applications]
      summary: Перевод заявки на другой этап
      description: |
        Переход проверяется по диаграмме этапов. Цели `TO_PRODUCTION` и `CLOSED`
        защищены: перейти в них обычным способом нельзя — только через проверку
        выпуска и через отдельное действие закрытия.
      parameters:
        - $ref: '#/components/parameters/Number'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [stage, lockVersion]
              properties:
                stage: { type: string }
                comment: { type: string }
                lossReasonCode: { type: string }
                lockVersion: { type: integer }
      responses:
        '200': { description: Этап изменён }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/applications/{number}/assign:
    post:
      tags: [applications]
      summary: Назначение исполнителя
      parameters:
        - $ref: '#/components/parameters/Number'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assigneeId]
              properties:
                assigneeId: { type: string }
                comment: { type: string }
                moveOpenTasks: { type: boolean, default: false }
      responses:
        '200': { description: Назначение выполнено, создана история и уведомления }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/applications/{number}/complexity:
    post:
      tags: [applications]
      summary: Подтверждение сложности
      parameters:
        - $ref: '#/components/parameters/Number'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [complexity, lockVersion]
              properties:
                complexity:
                  type: string
                  enum: [STANDARD, MODIFIED, CUSTOM, NEEDS_CLASSIFICATION]
                comment: { type: string }
                lockVersion: { type: integer }
      responses:
        '200': { description: Сложность подтверждена, заключения КО перепроверяются }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/applications/{number}/timeline:
    get:
      tags: [applications]
      summary: Лента событий заявки
      parameters:
        - $ref: '#/components/parameters/Number'
      responses:
        '200':
          description: События по убыванию времени
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TimelineEvent' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    integrationToken:
      type: apiKey
      in: header
      name: x-integration-token
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string }
      description: Повтор с тем же ключом не создаёт вторую запись.
    Number:
      name: number
      in: path
      required: true
      schema: { type: string }
      description: Номер заявки.
...
  responses:
    BadRequest:
      description: Ошибка запроса
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Требуется действительный токен
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Не найдено или вне области видимости
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Конфликт состояния или версии
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Нарушено бизнес-правило
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Превышен лимит частоты запросов
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Locked:
      description: Вход временно заблокирован после неудачных попыток
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [code, message, correlationId]
      properties:
        code: { type: string, example: VERSION_CONFLICT }
        message: { type: string }
        correlationId: { type: string }
        fields:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
        details:
          type: object
          additionalProperties: true
    User:
      type: object
      properties:
        id: { type: string }
        login: { type: string }
        fullName: { type: string }
        email: { type: string, nullable: true }
        role: { type: string }
        isActive: { type: boolean }
        blockedAt: { type: string, format: date-time, nullable: true }
    ApplicationLine:
      type: object
      properties:
        id: { type: string }
        lineId: { type: string }
        name: { type: string }
        quantity: { type: number }
        unit: { type: string }
        price: { type: number, nullable: true }
        complexity: { type: string, nullable: true }
        params:
          type: object
          additionalProperties: true
          description: |
            Параметры позиции. Признак `requiresKo` определяет обязательность
            заключения КО для MODIFIED и вычисляется на сервере.
    Application:
      type: object
      properties:
        id: { type: string }
        number: { type: string }
        stage: { type: string }
        complexity: { type: string }
        ownerId: { type: string }
        source: { type: string }
        lockVersion: { type: integer }
        versionNo: { type: integer }
        createdAt: { type: string, format: date-time }
        organization: { $ref: '#/components/schemas/Organization' }
        lines:
          type: array
          items: { $ref: '#/components/schemas/ApplicationLine' }
    ApplicationCreate:
      type: object
      required: [organizationName]
      properties:
        organizationName: { type: string }
        inn: { type: string, nullable: true }
        kpp: { type: string, nullable: true }
        contactName: { type: string, nullable: true }
        contactPhone: { type: string, nullable: true }
        contactEmail: { type: string, nullable: true }
        externalNumber: { type: string, nullable: true }
        source: { type: string }
        priority: { type: string, enum: [LOW, NORMAL, HIGH, URGENT] }
        lines:
          type: array
          items:
            type: object
            properties:
              lineId: { type: string }
              name: { type: string }
              quantity: { type: number }
              unit: { type: string }
              price: { type: number, nullable: true }
              complexity: { type: string, nullable: true }
    ApplicationUpdate:
      type: object
      required: [lockVersion]
      properties:
        lockVersion: { type: integer }
        contactName: { type: string }
        contactPhone: { type: string }
        contactEmail: { type: string }
        lines:
          type: array
          items: { $ref: '#/components/schemas/ApplicationLine' }
    ApplicationCreateResult:
      type: object
      properties:
        application: { $ref: '#/components/schemas/Application' }
        deduplicated: { type: boolean, description: true, если заявка уже существовала по этому ключу }
    PagedApplications:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Application' }
        total: { type: integer }
        page: { type: integer }
        size: { type: integer }
    Organization:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        inn: { type: string, nullable: true }
    TimelineEvent:
      type: object
      properties:
        at: { type: string, format: date-time }
        kind: { type: string }
        title: { type: string }
        actorName: { type: string, nullable: true }
        from: { type: string, nullable: true }
        to: { type: string, nullable: true }
...
security:
  - bearerAuth: []